AddPost 컴포넌트 생성, WritePost 컴포넌트 수정
